Switch to packer-based VM image build
The plan was to use EC2 AMI import after creating the raw VM image
locally. Unfortunately, the AMI import only works for a limited
set of known OSes and versions. The image build now uses packer's
Amazon ebssurrogate builder to create an EC2 instance and attach a
volume to it, onto which the container image filesystem is
extracted and then it is snapshotted to create the AMI.