Make `Security.RunAsUserID` in VMSpec work
This fixes a typo causing `Secutiry.RunAsUserID` to not be set if
defined in user data. It also changes the order of operations in
preinit so that setuid() comes last before running the entrypoint;
if it happens earlier and the new user does not have permission to
call setgid(), there would be an error.