Make preinit persistent by default
Initially the preinit process did setup and then exec()'d the
image entrypoint. Now it stays persistent by default so a reboot()
system call can be made when a poweroff signal is received. The
behavior to exec() the entrypoint can still be achieved by setting
the `replace-init` bool in user data. This may be useful if the
image bakes in its own init system.