Prevent unnecessary rebuild of mkfs.btrfs
A new build image is created and tagged on the first build after
a new commit. This can cause files to be updated that are newer
than mkfs.btrfs and cause btrfs-progs to be recompiled.

However, if there are no changes in btrfs-progs, a new binary of
mkfs.btrfs does not seem to be linked.

This updates the timestamp of mkfs.btrfs so this does not occur on
every subsequent build.