Add chronyd as a service
Instances will now start with chronyd running, in addition to the
command configured through a Dockerfile or overridden in user data.
Both processes will be configured within a new Service type that
has been added. A Supervisor has been added that starts and stops
services and waits for them to exit.