Add services and a supervisor
The main process and chrony are now started within a type
implementing a Service interface, which can start, stop, and kill
the process it manages.

The default service implementation restarts its process any time
it dies, whereas when the main process exits, it will cause all
processes to shut down and the machine to power off via a
Supervisor.

The Supervisor starts all services, including the main process, and
waits for the main process to exit or for a poweroff signal to be
received.