Add a login user for ssh
This adds a user before starting sshd, and retrieves the key from
instance metadata and writes it to the user's authorized_keys file.
It also makes the sshd service optional, so that if a key has not
been assigned to the instance, sshd will simply fail to start,
rather than throwing an error.