Remove shutdown hangs if processes have not started
Since the supervisor's shutdownAll() function waits for processes
to exit, it should be put into a goroutine when called from select
so it doesn't hang the case branch if processes do not exit.

In addition, a nil value should be put into a service's shutdown
channel if it receives a shutdown signal when no process is
running so a Wait() doesn't hang forever.