Enable shell and sudo for login user when ssh is enabled
This includes busybox in the image so the user will be sure to
have a shell, even if the container image does not inclue one. The
shell is actually a wrapper around the busybox shell that runs it
with --noprofile, so that the PATH compiled into sshd will be used
even if the container image has an /etc/profile. This ensures that
sudo is found within the /__cb__ directory.