Rename `ctr2ami`/`converter` to `ctr2disk`
There was an inconsistency in the naming, where the Go code
referred to `ctr2ami`, but it was used to produce an executable
called `converter`. The code is used to prepare a disk to create
an AMI but does not create the AMI itself. For this reason it has
been renamed to `ctr2disk` and references to `converter` have been
removed.