Enable running init scripts on boot
The user data may now contain a field `init-scripts`, which
contains an array of strings that must be scripts with a valid
interpreter. The scripts are run in the order defined in user data,
and have access to the environment of the main command. They are
run just before any services and the main command are executed.