Refactor build to use dependencies from separate repo
Instead of building the kernel and other dependencies in the easyto
repo, they are now built in easyto-assets. A Makefile.inc included
in the easyto Makfile is also built there that includes shared
variables and targets.

This also modifies the runtime files so that init.tar is built
here, but it now only includes the init from this repo written in
Go, and the old init.tar is now called base.tar and is built in
easyto-assets.